Add DISCONNECT event for OkCoin streaming
