+++ xchange-okcoin/src/main/java/com/xeiam/xchange/okcoin/service/streaming/WebSocketService.java
@@ -2,4 +2,5 @@ package com.xeiam.xchange.okcoin.service.streaming;
 
 public interface WebSocketService {
    public void onReceive(String msg);
+   public void onDisconnect();
 }
